css refactored
