+++ src/Components/GameFieldComponent/OverlayComponents/GameOverlays/ChildComponents/css/helpDialog.css
@@ -1,0 +1,169 @@
+.s-overlay-help
+{
+    width: 40%;
+    height: 70%;
+    background-color: rgb(32, 31, 31);
+    border-radius: 25px;
+    box-shadow: 0px 0px 50px silver;
+}
+.s-overlay-buttons {
+    display: flex;
+    width: 100%;
+    flex-direction: column;
+    align-items: center;
+}
+.s-overlay-button {
+    padding: 1% 5%;
+    font-family: "Montserrat";
+    font-weight: 600;
+    font-size: 2rem;
+    margin: 3% 0;
+    color: white;
+    background-color: darkorange;
+    border: 2px solid white;
+    border-radius: 5px;
+}
+.s-overlay-button:hover {
+    color: rgb(36, 35, 35);
+    border-color: rgb(255, 230, 0);
+    cursor: pointer;
+}
+.s-help-h {
+    text-align: center;
+    color: darkorange;
+    font-size: 4rem;
+    margin: 10% 0 5% 0;
+}
+.s-help-t {
+    color: white;
+    width: 90%;
+    padding: 3% 5%;
+    font-size: 2rem;
+}
+.s-help-t i {
+    color: red;
+}
+
+
+/* Extra small devices (phones, 600px and down) */
+@media only screen and (max-width: 600px) {
+    .s-overlay-button {
+        font-size: 0.75rem;
+        margin: 1.5% 0;
+    }
+    .s-overlay-help{
+        width: 50%;
+        height: 70%;
+    }
+    .s-help-h{
+        font-size: 2rem;
+        margin: 5% 0 2.5% 0;
+    }
+    .s-help-t {
+        padding: 3% 5%;
+        font-size: 0.75rem;
+    }
+}
+
+/* Small devices (portrait tablets and large phones, 600px and up) */
+@media only screen and (min-width: 600px) {
+
+    .s-overlay-button {
+        font-size: 0.75rem;
+        margin: 1.5% 0;
+    }
+    .s-overlay-help{
+        width: 50%;
+        height: 70%;
+    }
+    .s-help-h{
+        font-size: 2rem;
+        margin: 7.5% 0 2.5% 0;
+    }
+    .s-help-t {
+        padding: 4% 5%;
+        font-size: 0.75rem;
+    }
+}
+
+/* Medium devices (landscape tablets, 768px and up) */
+@media only screen and (min-width: 768px) {
+    .s-overlay-button {
+        font-size: 0.75rem;
+        margin: 1% 0;
+    }
+    .s-overlay-help{
+        width: 50%;
+        height: 70%;
+    }
+    .s-help-h{
+        font-size: 2rem;
+        margin: 7.5% 0 2.5% 0;
+    }
+    .s-help-t {
+        padding: 2% 5%;
+        font-size: 1rem;
+    }
+
+}
+
+/* Large devices (laptops/desktops, 992px and up) */
+@media only screen and (min-width: 992px) {
+    .s-overlay-button {
+        font-size: 1.25rem;
+        margin: 3% 0;
+    }
+    .s-overlay-help{
+        width: 40%;
+        height: 70%;
+    }
+    .s-help-h{
+        font-size: 3rem;
+        margin: 10% 0 5% 0;
+    }
+    .s-help-t {
+        padding: 3% 5%;
+        font-size: 1.25rem;
+    }
+}
+
+/* Extra large devices (large laptops and desktops, 1200px and up) */
+@media only screen and (min-width: 1200px) {
+
+    .s-overlay-button {
+        font-size: 2rem;
+        margin: 3% 0;
+    }
+    .s-overlay-help{
+        width: 40%;
+        height: 70%;
+    }
+    .s-help-h{
+        font-size: 4rem;
+        margin: 10% 0 5% 0;
+    }
+    .s-help-t {
+        padding: 3% 5%;
+        font-size: 1.5rem;
+    }
+}
+
+@media only screen and (min-width: 1600px) {
+    .s-overlay-button {
+        font-size: 2rem;
+        margin: 3% 0;
+    }
+    .s-overlay-menu,
+    .s-overlay-help{
+        width: 40%;
+        height: 70%;
+    }
+    .s-help-h{
+        font-size: 4rem;
+        margin: 10% 0 5% 0;
+    }
+    .s-help-t {
+        padding: 2% 5%;
+        font-size: 2rem;
+    }
+}
