+++ src/Components/GameFieldComponent/OverlayComponents/GameOverlays/css/menuOverlay.css
@@ -12,7 +12,7 @@
 }
 
 .overlay-box{
-    width: 30%;
+    width: 40%;
     height: 70%;
     background-color: rgb(32, 31, 31);
     border-radius: 25px;
@@ -20,9 +20,10 @@
 }
 .overlay-box h2{
     text-align: center;
-    padding: 10% 20%;
-    font-size: 300%;
-    color: white;
+    margin: 10% 0 5% 0;
+    font-size: 3rem;
+    color: darkorange;
+    font-weight: 400;
 }
 
 .overlay-controls{
@@ -40,52 +41,50 @@
 .hide-overlay{
     display: none;
 }
-
-.btn-game{
-    color: white;
+.s-overlay-button {
+    padding: 1% 5%;
     font-family: "Montserrat";
-    width: 50%;
-    background-color: orange;
-    border-radius: 25px;
-    border: 2px solid white;
-    font-size: 125%;
     font-weight: 600;
-    padding: 1.25% 2.5%;
-    transition: all 0.25s;
+    font-size: 2rem;
+    margin: 3% 0;
+    color: white;
+    background-color: darkorange;
+    border: 2px solid white;
+    border-radius: 5px;
 }
-.btn-game:hover{
+.s-overlay-button:hover {
     color: rgb(36, 35, 35);
     border-color: rgb(255, 230, 0);
     cursor: pointer;
 }
 
 
+
 @media only screen and (max-width: 600px) {
-    .btn-game{
-        border: 1px solid white;
-        font-size: 0.5rem;
-        margin-top: 4%;
+    .s-overlay-button{
+        border: 2px solid white;
+        font-size: 1rem;
+        margin-top: 2%;
     }
     .overlay-box h2{
         font-size: 1.5rem;
-        padding: 5% 20%;
+        margin: 10% 0 5% 0;
     }
     .overlay-box{
-        width: 50%;
+        width: 40%;
     }
-
 }
 
 /* Small devices (portrait tablets and large phones, 600px and up) */
 @media only screen and (min-width: 600px) {
-    .btn-game{
-        border: 1px solid white;
-        font-size: 0.7rem;
-        margin-top: 3%;
+    .s-overlay-button{
+        border: 2px solid white;
+        font-size: 1rem;
+        margin-top: 2%;
     }
     .overlay-box h2{
-        font-size: 1.5rem;
-        padding: 3% 20% 0% 20%;
+        font-size: 1.75rem;
+        margin: 10% 0 5% 0;
     }
     .overlay-box{
         width: 50%;
@@ -94,14 +93,14 @@
 
 /* Medium devices (landscape tablets, 768px and up) */
 @media only screen and (min-width: 768px){
-    .btn-game{
-        border: 1px solid white;
-        font-size: 0.9rem;
-        margin-top: 3%;
+    .s-overlay-button{
+        border: 2px solid white;
+        font-size: 1.25rem;
+        margin-top: 2%;
     }
     .overlay-box h2{
         font-size: 2rem;
-        padding: 3% 20%;
+        margin: 10% 0 5% 0;;
     }
     .overlay-box{
         width: 50%;
@@ -110,32 +109,64 @@
 
 /* Large devices (laptops/desktops, 992px and up) */
 @media only screen and (min-width: 992px) {
-    .btn-game{
+    .s-overlay-button{
         border: 2px solid white;
-        font-size: 1.5rem;
-        margin-top: 5%;
+        font-size: 1.75rem;
+        margin-top: 2.5%;
     }
     .overlay-box h2{
         font-size: 3rem;
-        padding: 5% 20%;
+        margin: 10% 0 5% 0;
+    }
+    .overlay-box{
+        width: 60%;
+    }
+}
+
+/* Extra large devices (large laptops and desktops, 1200px and up) */
+@media only screen and (min-width: 1200px) {
+    .s-overlay-button{
+        border: 2px solid white;
+        font-size: 1.75rem;
+        margin-top: 2%;
+    }
+    .overlay-box h2{
+        font-size: 4rem;
+        margin: 10% 0 5% 0;
+    }
+    .overlay-box{
+        width: 60%;
+    }
+}
+
+/* Extra large devices (large laptops and desktops, 1400px and up) */
+@media only screen and (min-width: 1400px) {
+    .s-overlay-button{
+        border: 2px solid white;
+        font-size: 2rem;
+        margin-top: 2.5%;
+    }
+    .overlay-box h2{
+        font-size: 4rem;
+        margin: 10% 0 5% 0;
     }
     .overlay-box{
         width: 40%;
     }
 }
 
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1200px) {
-    .btn-game{
+/* Extra large devices (large laptops and desktops, 1600px and up) */
+@media only screen and (min-width: 1600px) {
+    .s-overlay-button{
         border: 2px solid white;
-        font-size: 1.5rem;
-        margin-top: 7.5%;
+        font-size: 2.25rem;
+        margin-top: 2.5%;
     }
     .overlay-box h2{
         font-size: 4rem;
-        padding: 5% 20%;
+        margin: 10% 0 5% 0;
     }
     .overlay-box{
-        width: 45%;
+        width: 40%;
     }
 }
